Decompose App god struct into focused files
Split internal/tui/app.go (1365 lines) into 5 cohesive files:

- app.go (672 lines): struct definition, constructors, Init/Update/View,
  message routing, layout computation, main view rendering
- exec.go (138 lines): execution lifecycle — startExecution, runStreaming,
  runBlocking, emitDryRun, createLogFile, buildEnv
- mouse.go (253 lines): mouse coordinate translation, hit testing, focus
  targeting, click-to-copy handlers for metadata and output
- selection.go (179 lines): text selection highlighting, extraction from
  rendered view, border badge injection
- overlays.go (173 lines): overlay open/close (wizard, confirm, palette)
  and overlay view rendering, var resolution

All files remain in the internal/tui package. No behavior changes —
pure structural refactor. BubbleTea message flow is unchanged.
