+++ internal/tui/app.go
@@ -2,9 +2,7 @@ package tui
 
 import (
 	"context"
-	"fmt"
 	"image/color"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -16,18 +14,15 @@ import (
 	"dops/internal/theme"
 	"dops/internal/tui/confirm"
 	"dops/internal/tui/footer"
-	"dops/internal/tui/help"
 	"dops/internal/tui/metadata"
 	"dops/internal/tui/output"
 	"dops/internal/tui/palette"
 	"dops/internal/tui/sidebar"
 	"dops/internal/tui/wizard"
 	"dops/internal/update"
-	"dops/internal/vars"
 
 	tea "charm.land/bubbletea/v2"
 	lipgloss "charm.land/lipgloss/v2"
-	"github.com/charmbracelet/x/ansi"
 )
 
 type viewState int
@@ -151,11 +146,6 @@ func (m App) computeLayout() layoutDims {
 		gap:              gap,
 		borderSize:       borderSize,
 	}
-}
-
-type executionDoneMsg struct {
-	LogPath string
-	Err     error
 }
 
 // ProgramRef holds a reference to the tea.Program that can be set after
@@ -504,183 +494,6 @@ func (m App) routeToComponent(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return m, nil
 }
 
-func (m App) startExecution(rb domain.Runbook, cat domain.Catalog, params map[string]string) (tea.Model, tea.Cmd) {
-	m.output.Clear()
-	resolved := m.resolveVars()
-	cmdStr := wizard.BuildCommand(rb, params, resolved)
-	m.output.SetCommand(cmdStr)
-	m.resizeAll()
-
-	if m.deps.Runner == nil {
-		return m, nil
-	}
-
-	scriptPath := filepath.Join(adapters.ExpandHome(cat.RunbookRoot()), rb.Name, rb.Script)
-
-	if m.deps.DryRun {
-		m.emitDryRun(scriptPath, params)
-		return m, nil
-	}
-
-	env := buildEnv(params)
-	logPath := m.createLogFile(cat.Name, rb.Name)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	m.cancelExec = cancel
-	m.execRunning = true
-
-	// Stream lines live via p.Send() if we have a program reference,
-	// otherwise fall back to a tea.Cmd (tests).
-	if prog := m.program(); prog != nil {
-		go m.runStreaming(ctx, prog, scriptPath, env, logPath)
-		return m, nil
-	}
-
-	return m, m.runBlocking(ctx, scriptPath, env, logPath)
-}
-
-// emitDryRun writes dry-run output to the output pane without executing.
-func (m *App) emitDryRun(scriptPath string, params map[string]string) {
-	m.output, _ = m.output.Update(output.OutputLineMsg{Text: "[DRY RUN] Would execute:"})
-	m.output, _ = m.output.Update(output.OutputLineMsg{Text: fmt.Sprintf("  Script: %s", scriptPath)})
-	m.output, _ = m.output.Update(output.OutputLineMsg{Text: ""})
-	if len(params) > 0 {
-		m.output, _ = m.output.Update(output.OutputLineMsg{Text: "  Environment:"})
-		for k, v := range params {
-			m.output, _ = m.output.Update(output.OutputLineMsg{
-				Text: fmt.Sprintf("    %s=%s", strings.ToUpper(k), v),
-			})
-		}
-	}
-	m.output, _ = m.output.Update(output.ExecutionDoneMsg{})
-}
-
-// runStreaming executes a script in a goroutine, streaming lines to the
-// tea.Program via Send(). Used for live output in the TUI.
-func (m App) runStreaming(ctx context.Context, prog *tea.Program, scriptPath string, env map[string]string, logPath string) {
-	lw := m.deps.LogWriter
-	lines, errs := m.deps.Runner.Run(ctx, scriptPath, env)
-	for line := range lines {
-		if lw != nil {
-			lw.WriteLine(line.Text)
-		}
-		prog.Send(output.OutputLineMsg{Text: line.Text, IsStderr: line.IsStderr})
-	}
-	if lw != nil {
-		lw.Close()
-	}
-	prog.Send(executionDoneMsg{LogPath: logPath, Err: <-errs})
-}
-
-// runBlocking returns a tea.Cmd that executes a script synchronously.
-// Used in tests where no tea.Program reference is available.
-func (m App) runBlocking(ctx context.Context, scriptPath string, env map[string]string, logPath string) tea.Cmd {
-	runner := m.deps.Runner
-	lw := m.deps.LogWriter
-	return func() tea.Msg {
-		lines, errs := runner.Run(ctx, scriptPath, env)
-		for line := range lines {
-			if lw != nil {
-				lw.WriteLine(line.Text)
-			}
-		}
-		if lw != nil {
-			lw.Close()
-		}
-		return executionDoneMsg{LogPath: logPath, Err: <-errs}
-	}
-}
-
-// program returns the tea.Program reference, or nil if not set.
-func (m App) program() *tea.Program {
-	if m.deps.ProgramRef != nil {
-		return m.deps.ProgramRef.P
-	}
-	return nil
-}
-
-// createLogFile creates a log file for the execution and returns its path.
-func (m App) createLogFile(catalogName, runbookName string) string {
-	if m.deps.LogWriter == nil {
-		return ""
-	}
-	logPath, err := m.deps.LogWriter.Create(catalogName, runbookName, time.Now())
-	if err != nil {
-		return ""
-	}
-	return logPath
-}
-
-// buildEnv converts parameter keys to uppercase environment variable names.
-func buildEnv(params map[string]string) map[string]string {
-	env := make(map[string]string, len(params))
-	for k, v := range params {
-		env[strings.ToUpper(k)] = v
-	}
-	return env
-}
-
-func (m App) openPalette() (tea.Model, tea.Cmd) {
-	p := palette.New(m.width)
-	m.pal = &p
-	m.state = statePalette
-	return m, nil
-}
-
-func (m App) openWizard() (tea.Model, tea.Cmd) {
-	if m.selected == nil || m.selCat == nil {
-		return m, nil
-	}
-
-	resolved := m.resolveVars()
-
-	// If no parameters at all, go straight to confirm.
-	if len(m.selected.Parameters) == 0 {
-		return m.openConfirm(*m.selected, *m.selCat, resolved)
-	}
-
-	// Always show wizard — pre-fills saved values, user can accept or override.
-	wiz := wizard.New(wizard.WizardConfig{
-		Runbook:  *m.selected,
-		Catalog:  *m.selCat,
-		Resolved: resolved,
-	})
-	wiz.SetStyles(m.deps.Styles)
-	if m.deps.Config != nil {
-		wiz.SetStore(m.deps.Config, m.deps.Vault)
-	}
-	m.wizard = &wiz
-	m.state = stateWizard
-	return m, wiz.Init()
-}
-
-func (m App) openConfirm(rb domain.Runbook, cat domain.Catalog, params map[string]string) (tea.Model, tea.Cmd) {
-	// Low/Medium risk: skip confirmation, execute immediately.
-	if rb.RiskLevel == domain.RiskLow || rb.RiskLevel == domain.RiskMedium || rb.RiskLevel == "" {
-		return m.startExecution(rb, cat, params)
-	}
-	c := confirm.New(confirm.Params{
-		Runbook:  rb,
-		Catalog:  cat,
-		Resolved: params,
-		Width:    m.width * overlayWidthRatio / overlayWidthDenom,
-		Styles:   m.deps.Styles,
-	})
-	m.conf = &c
-	m.state = stateConfirm
-	return m, nil
-}
-
-func (m App) resolveVars() map[string]string {
-	if m.deps.Config == nil || m.selected == nil || m.selCat == nil {
-		return make(map[string]string)
-	}
-	// Vault stores all values as plaintext inside the encrypted blob,
-	// so no per-value decryption is needed.
-	resolver := vars.NewDefaultResolver()
-	return resolver.Resolve(m.deps.Config, m.selCat.Name, m.selected.Name, m.selected.Parameters)
-}
-
 func (m App) View() tea.View {
 	// Guard: before WindowSizeMsg arrives, width/height are defaults (80x24).
 	// In alt screen, BubbleTea sends WindowSizeMsg on startup, but View()
@@ -835,518 +648,12 @@ func (m App) viewNormal() tea.View {
 	return tea.NewView(content)
 }
 
-func (m App) viewHelpOverlay() tea.View {
-	var helpFocus help.FocusTarget
-	if m.focus == focusOutput {
-		helpFocus = help.FocusOutput
-	}
-	helpView := help.Render(helpFocus, m.width/2, m.deps.Styles)
-
-	content := lipgloss.Place(m.width, m.height,
-		lipgloss.Center, lipgloss.Center,
-		helpView,
-	)
-
-	footerView := footer.Render(footer.StateHelp, m.width, m.deps.Styles, "")
-	content = lipgloss.JoinVertical(lipgloss.Left, content, footerView)
-
-	return tea.NewView(content)
-}
-
-func (m App) viewConfirmOverlay() tea.View {
-	confView := m.conf.View()
-
-	overlayW := m.width * overlayWidthRatio / overlayWidthDenom
-	if overlayW < overlayMinWidth {
-		overlayW = overlayMinWidth
-	}
-
-	var primaryFg color.Color
-	primaryFg = lipgloss.NoColor{}
-	if m.deps.Styles != nil {
-		primaryFg = m.deps.Styles.Primary.GetForeground()
-	}
-
-	overlay := lipgloss.NewStyle().
-		Border(lipgloss.ThickBorder()).
-		BorderForeground(primaryFg).
-		Padding(1, 2).
-		Width(overlayW).
-		Render(confView)
-
-	content := lipgloss.Place(m.width, m.height,
-		lipgloss.Center, lipgloss.Center,
-		overlay,
-	)
-
-	footerView := footer.Render(footer.StateConfirm, m.width, m.deps.Styles, "")
-	content = lipgloss.JoinVertical(lipgloss.Left, content, footerView)
-
-	return tea.NewView(content)
-}
-
-func (m App) viewWizardOverlay() tea.View {
-	wizView := m.wizard.View()
-
-	overlayW := m.width * overlayWidthRatio / overlayWidthDenom
-	if overlayW < overlayMinWidth {
-		overlayW = overlayMinWidth
-	}
-
-	// Left accent bar only — transparent background (no Background() set).
-	var primaryFg color.Color
-	primaryFg = lipgloss.NoColor{}
-	if m.deps.Styles != nil {
-		primaryFg = m.deps.Styles.Primary.GetForeground()
-	}
-
-	overlay := lipgloss.NewStyle().
-		Border(lipgloss.ThickBorder()).
-		BorderForeground(primaryFg).
-		Padding(1, 2).
-		Width(overlayW).
-		Render(wizView)
-
-	content := lipgloss.Place(m.width, m.height,
-		lipgloss.Center, lipgloss.Center,
-		overlay,
-	)
-
-	footerView := footer.Render(footer.StateWizard, m.width, m.deps.Styles, "")
-	content = lipgloss.JoinVertical(lipgloss.Left, content, footerView)
-
-	return tea.NewView(content)
-}
-
-func (m App) viewPaletteOverlay() tea.View {
-	palView := m.pal.View()
-
-	overlay := lipgloss.NewStyle().
-		Width(m.width).
-		Render(palView)
-
-	footerView := footer.Render(footer.StatePalette, m.width, m.deps.Styles, "")
-	content := lipgloss.JoinVertical(lipgloss.Left, overlay, footerView)
-
-	return tea.NewView(content)
-}
-
 // clamp returns v if v >= min, otherwise min.
 func clamp(v, min int) int {
 	if v < min {
 		return min
 	}
 	return v
-}
-
-// translateMouseForOutput converts terminal-absolute mouse coordinates to
-// output content-relative coordinates (inside the output border).
-func (m App) translateMouseForOutput(msg tea.Msg) tea.Msg {
-	l := m.computeLayout()
-	originX := layoutMarginLeft + l.sidebarRenderedW + l.gap + layoutBorderSize
-	originY := layoutMarginTop + l.metaRenderedH + layoutBorderSize
-
-	switch msg := msg.(type) {
-	case tea.MouseClickMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg
-	case tea.MouseReleaseMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg
-	case tea.MouseMotionMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg
-	case tea.MouseWheelMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg
-	}
-	return msg
-}
-
-// outputPaneBounds computes the output pane's terminal-absolute bounds
-// for selection highlight and text extraction confinement.
-func (m App) outputPaneBounds() (top, bottom, left, right int) {
-	l := m.computeLayout()
-	padX := 1
-	logW := max(1, l.outputInnerW-padX*2-1) // -padX*2 for left/right pad, -1 for scrollbar
-
-	left = layoutMarginLeft + l.sidebarRenderedW + l.gap + layoutBorderSize + padX
-	right = left + logW
-	top = layoutMarginTop + l.metaRenderedH + layoutBorderSize
-	bottom = top + l.outputContentH
-	return
-}
-
-func (m App) extractSelectionFromView() string {
-	sel := m.output.Selection()
-	if !sel.Active || sel.IsEmpty() {
-		return ""
-	}
-
-	view := m.viewNormal()
-	content := view.Content
-	startX, startY, endX, endY := sel.Bounds()
-
-	// Confine extraction to output pane bounds.
-	bTop, bBottom, bLeft, bRight := m.outputPaneBounds()
-	if startY < bTop {
-		startY = bTop
-		startX = bLeft
-	}
-	if endY > bBottom {
-		endY = bBottom
-		endX = bRight
-	}
-
-	lines := strings.Split(content, "\n")
-	var result []string
-	for i := startY; i <= endY; i++ {
-		if i < 0 || i >= len(lines) {
-			continue
-		}
-		lineWidth := ansi.StringWidth(lines[i])
-		if lineWidth == 0 {
-			continue
-		}
-
-		lx := bLeft
-		rx := min(lineWidth, bRight)
-		if i == startY {
-			lx = max(startX, bLeft)
-		}
-		if i == endY {
-			rx = min(bRight, min(lineWidth, endX+1))
-		}
-		if lx >= rx {
-			continue
-		}
-
-		selected := ansi.Cut(lines[i], lx, rx)
-		plain := ansi.Strip(selected)
-		plain = strings.TrimRight(plain, " ")
-		if plain != "" {
-			result = append(result, plain)
-		}
-	}
-
-	return strings.TrimRight(strings.Join(result, "\n"), "\n ")
-}
-
-// selectionBounds defines the rectangular region of the output pane within
-// the full terminal view, used to clamp text selection highlighting.
-type selectionBounds struct {
-	top, bottom, left, right int
-}
-
-// applySelectionHighlight post-processes the full terminal view to highlight
-// the selected text, confined within the output pane bounds.
-func applySelectionHighlight(content string, sel output.TextSelection, styles *theme.Styles, bounds selectionBounds) string {
-	hlStyle := lipgloss.NewStyle()
-	if styles != nil {
-		hlStyle = lipgloss.NewStyle().
-			Background(styles.Primary.GetForeground()).
-			Foreground(styles.BackgroundElem.GetForeground())
-	}
-
-	startX, startY, endX, endY := sel.Bounds()
-
-	// Clamp to output pane bounds.
-	if startY < bounds.top {
-		startY = bounds.top
-		startX = bounds.left
-	}
-	if endY > bounds.bottom {
-		endY = bounds.bottom
-		endX = bounds.right
-	}
-	if startY > endY {
-		return content
-	}
-
-	lines := strings.Split(content, "\n")
-	for i := range lines {
-		lines[i] = highlightLine(lines[i], i, startX, startY, endX, endY, bounds, hlStyle)
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-// highlightLine applies selection highlighting to a single line if it falls
-// within the selection range. It returns the line unchanged when outside the
-// range or when the line has zero visible width.
-func highlightLine(line string, i, startX, startY, endX, endY int, bounds selectionBounds, hlStyle lipgloss.Style) string {
-	if i < startY || i > endY {
-		return line
-	}
-
-	lineWidth := ansi.StringWidth(line)
-	if lineWidth == 0 {
-		return line
-	}
-
-	var lx, rx int
-	if i == startY && i == endY {
-		lx = max(startX, bounds.left)
-		rx = min(bounds.right, min(lineWidth, endX+1))
-	} else if i == startY {
-		lx = max(startX, bounds.left)
-		rx = min(lineWidth, bounds.right)
-	} else if i == endY {
-		lx = bounds.left
-		rx = min(bounds.right, min(lineWidth, endX+1))
-	} else {
-		lx = bounds.left
-		rx = min(lineWidth, bounds.right)
-	}
-
-	if lx >= rx || lx >= lineWidth {
-		return line
-	}
-
-	before := ansi.Cut(line, 0, lx)
-	selected := ansi.Cut(line, lx, rx)
-	after := ansi.Cut(line, rx, lineWidth)
-
-	plain := ansi.Strip(selected)
-	return before + "\x1b[0m" + hlStyle.Render(plain) + after
-}
-
-// injectBorderBadge replaces part of the top border row with a styled badge.
-// The badge appears near the right end: ╭─── Copied to Clipboard! ──╮
-func injectBorderBadge(rendered, text string, styles *theme.Styles) string {
-	lines := strings.Split(rendered, "\n")
-	if len(lines) == 0 {
-		return rendered
-	}
-
-	topLine := lines[0]
-	topWidth := lipgloss.Width(topLine)
-
-	// Style the badge text.
-	badgeStyle := lipgloss.NewStyle()
-	if styles != nil {
-		badgeStyle = styles.Success
-	}
-	badge := " " + badgeStyle.Render(text) + " "
-	badgeW := lipgloss.Width(badge)
-
-	// Place badge near the right end (2 chars from right for the corner + margin).
-	rightMargin := 3
-	insertAt := topWidth - badgeW - rightMargin
-	if insertAt < 2 {
-		return rendered // not enough space
-	}
-
-	// Rebuild the top line: keep left portion, insert badge, keep right portion.
-	left := ansi.Cut(topLine, 0, insertAt)
-	right := ansi.Cut(topLine, insertAt+badgeW, topWidth)
-	lines[0] = left + badge + right
-
-	return strings.Join(lines, "\n")
-}
-
-
-func sidebarWidth(totalWidth int) int {
-	w := totalWidth / 3
-	if w < sidebarMinWidth {
-		w = sidebarMinWidth
-	}
-	if w > sidebarMaxWidth {
-		w = sidebarMaxWidth
-	}
-	return w
-}
-
-// sidebarBounds returns the origin and whether coordinates are within the sidebar.
-func (m App) sidebarBounds(mx, my int) (originX, originY int, inBounds bool) {
-	originY = layoutMarginTop + layoutBorderSize
-	originX = layoutMarginLeft + layoutBorderSize + layoutPadLeft
-	innerW := clamp(m.width-layoutMarginLeft, 1)
-	sidebarW := sidebarWidth(innerW) + layoutBorderSize*2 + layoutPadLeft
-	inBounds = mx >= layoutMarginLeft && mx < layoutMarginLeft+sidebarW &&
-		my >= layoutMarginTop && my < m.height
-	return
-}
-
-// translateMouseForSidebar converts terminal-absolute mouse coordinates to
-// sidebar content-relative coordinates. Returns translated msg and inBounds.
-func (m App) translateMouseForSidebar(msg tea.Msg) (tea.Msg, bool) {
-	mx, my, ok := mouseCoords(msg)
-	if !ok {
-		return msg, false
-	}
-	originX, originY, inBounds := m.sidebarBounds(mx, my)
-
-	switch msg := msg.(type) {
-	case tea.MouseClickMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg, inBounds
-	case tea.MouseMotionMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg, inBounds
-	case tea.MouseReleaseMsg:
-		msg.X -= originX
-		msg.Y -= originY
-		return msg, inBounds
-	case tea.MouseWheelMsg:
-		return msg, inBounds
-	}
-	return msg, false
-}
-
-// mouseCoords extracts X, Y from any mouse message type.
-// Returns (0, 0, false) if msg is not a mouse event.
-func mouseCoords(msg tea.Msg) (x, y int, ok bool) {
-	switch msg := msg.(type) {
-	case tea.MouseClickMsg:
-		return msg.X, msg.Y, true
-	case tea.MouseReleaseMsg:
-		return msg.X, msg.Y, true
-	case tea.MouseMotionMsg:
-		return msg.X, msg.Y, true
-	case tea.MouseWheelMsg:
-		return msg.X, msg.Y, true
-	}
-	return 0, 0, false
-}
-
-// focusTargetFromMouse returns which pane a mouse event is over.
-func (m App) focusTargetFromMouse(msg tea.Msg) (focusTarget, bool) {
-	mx, my, ok := mouseCoords(msg)
-	if !ok {
-		return 0, false
-	}
-
-	innerW := clamp(m.width-layoutMarginLeft, 1)
-	sidebarW := sidebarWidth(innerW) + layoutBorderSize*2 + layoutPadLeft
-	sidebarRight := layoutMarginLeft + sidebarW
-
-	if mx >= layoutMarginLeft && mx < sidebarRight && my >= layoutMarginTop {
-		// Sidebar: only steal focus on click, not hover.
-		if _, isClick := msg.(tea.MouseClickMsg); isClick {
-			return focusSidebar, true
-		}
-		return m.focus, false
-	}
-	if mx >= sidebarRight && my >= layoutMarginTop {
-		return focusOutput, true
-	}
-	return m.focus, false
-}
-
-func isMouseMsg(msg tea.Msg) bool {
-	switch msg.(type) {
-	case tea.MouseClickMsg, tea.MouseMotionMsg, tea.MouseReleaseMsg, tea.MouseWheelMsg:
-		return true
-	}
-	return false
-}
-
-func isMouseClick(msg tea.Msg) bool {
-	_, ok := msg.(tea.MouseClickMsg)
-	return ok
-}
-
-// handleMetadataClick checks if a mouse click landed on the path/URL line
-// in the metadata panel. If so, returns the location string and a cmd to copy it.
-func (m App) handleMetadataClick(msg tea.Msg) (string, tea.Cmd) {
-	click, ok := msg.(tea.MouseClickMsg)
-	if !ok {
-		return "", nil
-	}
-
-	location := metadata.Location(m.selected, m.selCat)
-	if location == "" {
-		return "", nil
-	}
-
-	l := m.computeLayout()
-
-	// Path is the last content line before the bottom border.
-	pathLineY := layoutMarginTop + l.metaRenderedH - layoutBorderSize - 1
-	if click.Y != pathLineY {
-		return "", nil
-	}
-
-	pathXStart := layoutMarginLeft + l.sidebarRenderedW + l.gap + layoutBorderSize + 1
-	pathXEnd := pathXStart + len(location)
-	if click.X < pathXStart || click.X >= pathXEnd {
-		return "", nil
-	}
-
-	return location, tea.Batch(
-		tea.SetClipboard(location),
-		tea.Tick(copyFlashDuration, func(time.Time) tea.Msg {
-			return copiedFlashMsg{}
-		}),
-	)
-}
-
-// handleOutputClick checks if a mouse click landed on the output header (command)
-// or footer (log path) text. If so, copies the value to clipboard.
-func (m *App) handleOutputClick(msg tea.Msg) tea.Cmd {
-	click, ok := msg.(tea.MouseClickMsg)
-	if !ok {
-		return nil
-	}
-
-	lines := strings.Split(ansi.Strip(m.viewNormal().Content), "\n")
-	if click.Y < 0 || click.Y >= len(lines) {
-		return nil
-	}
-	line := lines[click.Y]
-
-	var copyText, region string
-
-	// Check footer FIRST — specific text match on "Saved to <path>".
-	if logPath := m.output.LogPath(); logPath != "" {
-		if strings.Contains(line, "Saved to") || strings.Contains(line, logPath) {
-			copyText, region = logPath, "footer"
-		}
-	}
-
-	// Check header — any line containing "$ " or "--param" in the output area.
-	if region == "" {
-		if cmd := m.output.Command(); cmd != "" {
-			if strings.Contains(line, "$ ") || strings.Contains(line, "--param") {
-				copyText, region = cmd, "header"
-			}
-		}
-	}
-
-	if region == "" {
-		return nil
-	}
-
-	// Guard: reject if a copy is already in progress.
-	if !m.output.TryCopy() {
-		return nil
-	}
-
-	// Green flash on the clicked region.
-	switch region {
-	case "header":
-		m.output.SetCopiedHeader(true)
-	case "footer":
-		m.output.SetCopiedFooter(true)
-	}
-	return tea.Batch(
-		tea.SetClipboard(copyText),
-		// Short highlight flash (500ms).
-		tea.Tick(copyFlashDuration, func(time.Time) tea.Msg {
-			return output.CopiedRegionFlashMsg{}
-		}),
-		// Badge stays longer (1.5s).
-		tea.Tick(copyFlashDuration, func(time.Time) tea.Msg {
-			return output.CopyFlashExpiredMsg{}
-		}),
-	)
 }
 
 func appFooterState(s viewState) footer.State {
